+++ xata/app/entry.server.tsx
@@ -1,7 +1,6 @@
 import type { EntryContext } from '@remix-run/node'
 import { RemixServer } from '@remix-run/react'
 import { renderToString } from 'react-dom/server'
-import { SECURITY_HEADERS } from './lib/settings'
 
 export default function handleRequest(
   request: Request,
@@ -14,10 +13,6 @@ export default function handleRequest(
   )
 
   responseHeaders.set('Content-Type', 'text/html')
-
-  Object.entries(SECURITY_HEADERS).forEach(([key, value]) => {
-    responseHeaders.set(key, value)
-  })
 
   return new Response('<!DOCTYPE html>' + markup, {
     status: responseStatusCode,
